backend/requirements.txt
aiosqlite==0.22.1
fastapi==0.135.1
httpx==0.28.1
pydantic==2.12.5
python-dotenv==1.2.2
uvicorn==0.41.0
